requirements.txt
python-telegram-bot==20.0
SQLAlchemy
psycopg2-binary
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.8
logging==0.4.9.6
python-telegram-bot==20.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
urllib3==2.2.2
